requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
alembic==1.4.2            # via flask-migrate
babel==2.8.0              # via flask-babelex
bcrypt==3.1.7             # via -r requirements.in, flask-user
blinker==1.4              # via flask-mail
cffi==1.14.0              # via bcrypt, cryptography
click==7.1.2              # via -r requirements.in, flask
cryptography==2.9.2       # via flask-user
dnspython==1.16.0         # via email-validator
email-validator==1.1.1    # via -r requirements.in
flask-babelex==0.9.4      # via -r requirements.in
flask-login==0.5.0        # via flask-user
flask-mail==0.9.1         # via flask-user
flask-migrate==2.5.3      # via -r requirements.in
flask-script==2.0.6       # via -r requirements.in
flask-sqlalchemy==2.4.3   # via -r requirements.in, flask-migrate, flask-user
flask-user==1.0.2.2       # via -r requirements.in
flask-wtf==0.14.3         # via flask-user
flask==1.1.2              # via -r requirements.in, flask-babelex, flask-login, flask-mail, flask-migrate, flask-script, flask-sqlalchemy, flask-user, flask-wtf
gunicorn==20.0.4          # via -r requirements.in
idna==2.9                 # via email-validator
itsdangerous==1.1.0       # via flask, flask-wtf
jinja2==2.11.2            # via flask, flask-babelex
mako==1.1.3               # via alembic
markupsafe==1.1.1         # via jinja2, mako, wtforms
passlib==1.7.2            # via -r requirements.in, flask-user
psycopg2==2.8.5           # via -r requirements.in
pycparser==2.20           # via cffi
python-dateutil==2.8.1    # via alembic
python-dotenv==0.13.0     # via -r requirements.in
python-editor==1.0.4      # via alembic
pytz==2020.1              # via babel
safe==0.4                 # via -r requirements.in
six==1.15.0               # via bcrypt, cryptography, python-dateutil, sqlalchemy-utils
speaklater==1.3           # via flask-babelex
sqlalchemy-utils==0.36.6  # via -r requirements.in
sqlalchemy==1.3.17        # via alembic, flask-sqlalchemy, sqlalchemy-utils
werkzeug==1.0.1           # via flask
wtforms==2.3.1            # via flask-wtf

# The following packages are considered to be unsafe in a requirements file:
# setuptools
